requirements.txt
Flask
requests
PyPDF2
Pillow
pytesseract
pdf2image
python-dotenv